Moved main functions to the Addons folder.
