+++ SparseGrids/tsgCoreOneDimensional.hpp
@@ -197,28 +197,6 @@ namespace OneDimensionalMeta{
 }
 
 //! \internal
-//! \brief Contains algorithms for generating exotic quadrature rules.
-//! \ingroup TasmanianCoreOneDimensional
-namespace OneDimensionalExoticQuad {
-    //! \brief Generate all roots up to the n roots of the n-th degree orthogonal polynomial.
-    std::vector<std::vector<double>> getRootCache(
-        const int n,
-        const std::vector<double> &ref_integral_weights,
-        const std::vector<double> &ref_points);
-    //! \brief Generate the full cache of Exotic (weighted and possibly shifted)
-    //! Gauss-Legendre weights and points for an (input) max level \b n, using
-    //! parameters \b shift, \b weight_fn, and \b nref. By default, the reference
-    //! points and reference weights are generated by the Gauss-Legendre rule.
-    void getExoticGaussLegendreCache(
-        const int n,
-        const double shift,
-        std::function<double(double)> weight_fn,
-        const int nref,
-        std::vector<std::vector<double>> &weights_cache,
-        std::vector<std::vector<double>> &points_cache);
-};
-
-//! \internal
 //! \brief Contains algorithms for generating Gauss, Chebyshev, and Fourier nodes and weights.
 //! \ingroup TasmanianCoreOneDimensional
 namespace OneDimensionalNodes{
